+++ nf_core/modules/lint/__init__.py
@@ -9,6 +9,7 @@ nf-core modules lint
 
 from __future__ import print_function
 import logging
+from nf_core.modules.modules_command import ModuleCommand
 import operator
 import os
 import questionary
@@ -16,6 +17,7 @@ import re
 import requests
 import rich
 import yaml
+import json
 from rich.console import Console
 from rich.table import Table
 from rich.markdown import Markdown
@@ -28,7 +30,6 @@ import nf_core.utils
 import nf_core.modules.module_utils
 from nf_core.modules.modules_repo import ModulesRepo
 from nf_core.modules.nfcore_module import NFCoreModule
-
 
 log = logging.getLogger(__name__)
 
@@ -50,7 +51,7 @@ class LintResult(object):
         self.module_name = mod.module_name
 
 
-class ModuleLint(object):
+class ModuleLint(ModuleCommand):
     """
     An object for linting modules either in a clone of the 'nf-core/modules'
     repository or in any nf-core pipeline directory
@@ -63,6 +64,7 @@ class ModuleLint(object):
     from .module_changes import module_changes
     from .module_tests import module_tests
     from .module_todos import module_todos
+    from .module_version import module_version
 
     def __init__(self, dir, key=()):
         self.dir = dir
@@ -78,10 +80,15 @@ class ModuleLint(object):
         self.lint_tests = ["main_nf", "functions_nf", "meta_yml", "module_changes", "module_todos"]
         self.key = key
         self.lint_config = None
+        self.modules_json = None
 
-        # Add tests specific to nf-core/modules
+        # Add tests specific to nf-core/modules or pipelines
         if self.repo_type == "modules":
             self.lint_tests.append("module_tests")
+
+        if self.repo_type == "pipeline":
+            # Add as first test to load git_sha before module_changes
+            self.lint_tests.insert(0, "module_version")
 
     def lint(self, module=None, all_modules=False, print_results=True, show_passed=False, local=False):
         """
@@ -156,9 +163,10 @@ class ModuleLint(object):
             log.info("Only running tests: '{}'".format("', '".join(self.key)))
             self.lint_tests = [k for k in self.lint_tests if k in self.key]
 
-        # If it is a pipeline, load the lint config file
+        # If it is a pipeline, load the lint config file and the modules.json file
         if self.repo_type == "pipeline":
-            self._load_lint_config()
+            self.load_lint_config()
+            self.modules_json = self.load_modules_json()
 
             # Only continue if a lint config has been loaded
             if self.lint_config:
@@ -410,25 +418,3 @@ class ModuleLint(object):
             self.passed += [LintResult(mod, m[0], m[1], m[2]) for m in mod.passed]
             self.warned += [LintResult(mod, m[0], m[1], m[2]) for m in mod.warned]
             self.failed += [LintResult(mod, m[0], m[1], m[2]) for m in mod.failed]
-
-    def _load_lint_config(self):
-        """Parse a pipeline lint config file.
-
-        Look for a file called either `.nf-core-lint.yml` or
-        `.nf-core-lint.yaml` in the pipeline root directory and parse it.
-        (`.yml` takes precedence).
-
-        Add parsed config to the `self.lint_config` class attribute.
-        """
-        config_fn = os.path.join(self.dir, ".nf-core-lint.yml")
-
-        # Pick up the file if it's .yaml instead of .yml
-        if not os.path.isfile(config_fn):
-            config_fn = os.path.join(self.dir, ".nf-core-lint.yaml")
-
-        # Load the YAML
-        try:
-            with open(config_fn, "r") as fh:
-                self.lint_config = yaml.safe_load(fh)
-        except FileNotFoundError:
-            log.debug("No lint config file found: {}".format(config_fn))