+++ nf_core/modules/lint/module_tests.py
@@ -1,3 +1,6 @@
+"""
+Lint the tests of a module in nf-core/modules
+"""
 import os
 import logging
 import sys